List Retrieval from Database
Using JQuery's AJAX to query the database and retrieve info (Checklist
items and ID in this case).

The function calls the 'test.php' file. Through php, a connection to
the DB is made, information is retrieved and is then encoded into an
JSON object and echoed.

The JQuery AJAX call receives the object, and in the 'success' field,
there's a callback function handling the data. In this case, it appends
the item and id into the list defined below in checklist.php

***Tested with names and ages initially and never changed Age in the
database, so ID in this code is Age.