Changes Made - See Description
**Implemented the designer view of the checklist into the Express app.
**Modified app.'s and index.js to handle the designer view routing.
**In addition to implementing routing support, a new boolean database
entry was added--called "admin"--to differentiate between regular users
and admins. By default it is set to "false" when a user registers. If
you want to test it, either change admin:false to true in the code, or
manually add an entry into MongoDB through the shell.
**Modified settingUpdate() to require the user to input strings greater
than length 5. Failure to do so results in a message being flashed, and
a redirect to the settings page. The function also makes sure that
users' passwords match before updating the database, failure of that
also results in a flash message.

**Might've forgotten some changes. They can all be seen on here though
if you go through